Fix BLE_Beacon demo and modify readme
Modify the BLE_Beacon demo to allocate the iBeacon object in heap rather than
in the stack where it is lost after the function terminates. Also, modified
BLE_EddystoneService readme file to include instructions on switching
SoftDevice when building for 16K targets. Finally changes the module name of
the BLE_EddystoneService in the module.json file.